requirements.txt
more-itertools==8.13.0
pycoax @ git+https://github.com/hanshuebner/coax@interface3#subdirectory=pycoax
pytn3270 @ git+https://github.com/hanshuebner/pytn3270
sliplib==0.6.2
sortedcontainers==2.4.0
telnetlib3==2.0.4
wcwidth==0.2.5
